requirements.txt
pandas
pint
pyyaml